Added StlObject class for importing Stl files
Added new StlObject class which is used to load external Stl files.

Fixed memory issue when adding a large number of triangles to 3d objects. The vectors used for vertex and triangle storage were being resized, causing vertex pointers to become invalid. Reserving enough space in the vectors before building the geometry solved this problem.

Added static triangles to Object class. These triangles are not affected by transformations.

Changed all double coordinates to floats to cut down on memory usage. There is no noticeable change in image quality.

+++ include/Primitives.hpp
@@ -20,7 +20,7 @@ namespace Primitives {
 		{
 		}
 
-		Plane(const vector3& pos_, const double& X, const double& Y) :
+		Plane(const vector3& pos_, const float& X, const float& Y) :
 			object(pos_),
 			dX(X/2),
 			dY(Y/2)
@@ -32,8 +32,8 @@ namespace Primitives {
 		virtual void userBuild();
 
 	protected:
-		double dX; ///< Half width of plane along the x-axis
-		double dY; ///< Half height of plane along the y-axis
+		float dX; ///< Half width of plane along the x-axis
+		float dY; ///< Half height of plane along the y-axis
 	};
 
 	class Cube : public Plane {
@@ -50,7 +50,7 @@ namespace Primitives {
 		{
 		}
 
-		Cube(const vector3& pos_, const double& X, const double& Y, const double& Z) :
+		Cube(const vector3& pos_, const float& X, const float& Y, const float& Z) :
 			Plane(pos_, X, Y),
 			dZ(Z/2)
 		{
@@ -61,7 +61,7 @@ namespace Primitives {
 		virtual void userBuild();
 
 	protected:
-		double dZ; ///< Half depth of cube along the z-axis
+		float dZ; ///< Half depth of cube along the z-axis
 	};
 
 	class Circle : public object {
@@ -80,7 +80,7 @@ namespace Primitives {
 		{
 		}
 
-		Circle(const vector3& pos_, const double& R, const int& sides) :
+		Circle(const vector3& pos_, const float& R, const int& sides) :
 			object(pos_),
 			dR(R),
 			nVertices(sides)
@@ -92,11 +92,11 @@ namespace Primitives {
 		virtual void userBuild();
 
 	protected:
-		double dR; ///< Radius of the circle
+		float dR; ///< Radius of the circle
 
 		int nVertices; ///< Number of vertices used to approximate the circumference of the circle (not counting the central vertex)
 
-		void approximate(const double& zoffset=0);
+		void approximate(const float& zoffset=0);
 	};
 
 	class Cylinder : public Circle {
@@ -113,7 +113,7 @@ namespace Primitives {
 		{
 		}
 
-		Cylinder(const vector3& pos_, const double& R, const double& Z, const int& sides) :
+		Cylinder(const vector3& pos_, const float& R, const float& Z, const int& sides) :
 			Circle(pos_, R, sides),
 			dZ(Z/2)
 		{
@@ -124,7 +124,7 @@ namespace Primitives {
 		virtual void userBuild();
 
 	protected:
-		double dZ; ///< Half depth of the cylinder along the x-axis
+		float dZ; ///< Half depth of the cylinder along the x-axis
 	};
 
 	class Cone : public Circle {
@@ -141,7 +141,7 @@ namespace Primitives {
 		{
 		}
 
-		Cone(const vector3& pos_, const double& R, const double& Z, const int& sides) :
+		Cone(const vector3& pos_, const float& R, const float& Z, const int& sides) :
 			Circle(pos_, R, sides),
 			dZ(Z / 2)
 		{
@@ -152,7 +152,7 @@ namespace Primitives {
 		virtual void userBuild();
 
 	private:
-		double dZ; ///< Half height of the cylinder along the z-axis
+		float dZ; ///< Half height of the cylinder along the z-axis
 	};
 
 	class Sphere : public object {
@@ -164,7 +164,7 @@ namespace Primitives {
 		virtual void userBuild();
 
 	protected:
-		double dR;
+		float dR;
 	};
 
 } // namespace primitives
