+++ include/plane.hpp
@@ -15,7 +15,11 @@ public:
 
 	/** Position and normal constructor
 	  */
-	plane(const vector3 &p_, const vector3 &n_) : p(p_), norm(n_) { }
+	plane(const vector3 &p_, const vector3 &n_) : 
+		p(p_), 
+		norm(n_) 
+	{ 
+	}
 
 	/** Constructor taking three points which lie on the surface of the plane
 	  */
@@ -25,14 +29,14 @@ public:
 	  * @param r The ray to check for intersection
 	  * @return True if the plane is in front of the ray, and return false if the plane is behind the ray (i.e. they do not intersect)
 	  */
-	bool intersects(const ray &r) const ;
+	bool intersects(const ray& r) const ;
 
 	/** Check if a ray intersects with the infinite plane defined by this object
 	  * @param r The ray to check for intersection
 	  * @param t The "distance" from the ray's vertex point, along its direction, to the point of intersection on the plane
 	  * @return True if the plane is in front of the ray, and return false if the plane is behind the ray (i.e. they do not intersect)
 	  */
-	bool intersects(const ray &r, double &t) const ;
+	bool intersects(const ray& r, float& t) const ;
 
 	/** Dump plane parameters to stdout
 	  */
