+++ source/camera.cpp
@@ -8,8 +8,8 @@
 
 const vector3 upVector(0, 1, 0);
 
-double WrappedValue::wrap(const double& d) const {
-	double retval = value + d;
+float WrappedValue::wrap(const float& d) const {
+	float retval = value + d;
 	if (!lock) {
 		retval = fmod((value + d) - minVal, delta);
 		if (retval < 0) {
@@ -29,8 +29,8 @@ double WrappedValue::wrap(const double& d) const {
 
 camera::camera() :
 	fov(pi/2), // radians
-	L(50E-3), // m
-	A(640.0/480), 
+	L(50E-3f), // m
+	A(640.0f/480), 
 	W(0),
 	H(0),
 	vPlane(),
@@ -62,17 +62,17 @@ camera::camera(const vector3 &pos_, const vector3 &dir_) :
 camera::~camera(){
 }
 
-void camera::setFOV(const double &fov_){ 
+void camera::setFOV(const float &fov_){ 
 	fov = fov_*deg2rad; 
 	computeViewingPlane();
 }
 
-void camera::setFocalLength(const double &length){
-	L = length*1E-3;
+void camera::setFocalLength(const float &length){
+	L = length*1E-3f;
 	computeViewingPlane();
 }
 
-void camera::setAspectRatio(const double &ratio){
+void camera::setAspectRatio(const float &ratio){
 	A = ratio;
 	computeViewingPlane();
 }
@@ -81,17 +81,17 @@ void camera::setAspectRatio(const double &ratio){
 // Movement methods
 /////////////////////////////////////////////////
 
-void camera::moveForward(const double &dist){
+void camera::moveForward(const float &dist){
 	pos += uZ*dist;
 	updateViewingPlane();
 }
 
-void camera::moveRight(const double &dist){
+void camera::moveRight(const float&dist){
 	pos += uX*dist;
 	updateViewingPlane();
 }
 
-void camera::moveUp(const double &dist){
+void camera::moveUp(const float&dist){
 	pos += uY*dist;
 	updateViewingPlane();	
 }
@@ -101,7 +101,7 @@ void camera::move(const vector3 &displacement){
 	updateViewingPlane();
 }
 
-void camera::move(const double &x, const double &y, const double &z){
+void camera::move(const float&x, const float&y, const float&z){
 	pos += vector3(x, y, z);
 	updateViewingPlane();
 }
@@ -111,7 +111,7 @@ void camera::moveTo(const vector3 &position){
 	updateViewingPlane();
 }
 
-void camera::moveTo(const double &x, const double &y, const double &z){
+void camera::moveTo(const float&x, const float&y, const float&z){
 	pos = vector3(x, y, z);
 	updateViewingPlane();
 }
@@ -120,7 +120,7 @@ void camera::moveTo(const double &x, const double &y, const double &z){
 // Rotation methods
 /////////////////////////////////////////////////
 
-void camera::rotate(const double &pitch, const double &yaw, const double& roll/*=0*/){
+void camera::rotate(const float& pitch, const float& yaw, const float& roll/*=0*/){
 	// Reset the camera to its original orientation
 	resetUnitVectors();
 	pitchAngle += pitch;
@@ -129,7 +129,7 @@ void camera::rotate(const double &pitch, const double &yaw, const double& roll/*
 	setRotation(pitchAngle.get(), rollAngle.get(), yawAngle.get());
 }
 
-void camera::setRotation(const double &theta, const double &phi, const double &psi){
+void camera::setRotation(const float& theta, const float& phi, const float& psi){
 	// Reset the camera to its original orientation
 	resetUnitVectors();
 	
@@ -173,7 +173,7 @@ void camera::resetOrientation(){
 // Rendering methods
 /////////////////////////////////////////////////
 
-void camera::render(const vector3 &offset, const triangle &tri, double *sX, double *sY, bool *valid){
+void camera::render(const vector3 &offset, const triangle &tri, float* sX, float* sY, bool* valid){
 	valid[0] = projectPoint(*tri.p0+offset, sX[0], sY[0]);
 	valid[1] = projectPoint(*tri.p1+offset, sX[1], sY[1]);
 	valid[2] = projectPoint(*tri.p2+offset, sX[2], sY[2]);
@@ -184,10 +184,10 @@ bool camera::checkCulling(const vector3 &offset, const triangle &tri){
 	return (((tri.p+offset) - pos) * tri.norm <= 0);
 }
 
-bool camera::projectPoint(const vector3 &vertex, double &sX, double &sY){
+bool camera::projectPoint(const vector3 &vertex, float& sX, float& sY){
 	ray proj(vertex, pos-vertex);
 	
-	double t;
+	float t;
 	if(vPlane.intersects(proj, t)){
 		// Get the point at which the ray intersect the viewing plane
 		vector3 pp = proj.extend(t) - vPlane.p;
@@ -239,17 +239,17 @@ void camera::resetUnitVectors() {
 	uZ = vector3(0, 0, 1);
 }
 
-void camera::convertToScreenSpace(const vector3 &vec, double &x, double &y){
+void camera::convertToScreenSpace(const vector3 &vec, float& x, float& y){
 	x = 2*(vec * uX)/W; // unitless
 	y = 2*(vec * uY)/H; // unitless
 }
 
-bool camera::rayTrace(const double &sX, const double &sY, const triangle &tri, vector3 &P){
-	double x = (sX * W)/2;
-	double y = (sY * H)/2;
+bool camera::rayTrace(const float& sX, const float& sY, const triangle &tri, vector3 &P){
+	float x = (sX * W)/2;
+	float y = (sY * H)/2;
 	vector3 vec = (vPlane.p + uX*x + uY*y - pos).normalize();
 	ray cast(pos, vec);
-	double t;
+	float t;
 	if(tri.intersects(cast, t)){
 		P = cast.extend(t);
 		return true;
