+++ source/plane.cpp
@@ -8,17 +8,17 @@ plane::plane(const vector3 &p0_, const vector3 &p1_, const vector3 &p2_){
 	
 	// Compute the normal to the plane
 	norm = p2p0.cross(p1p0).normalize();
-	p = (p0_ + p1_ + p2_)*(1/3.0);
+	p = (p0_ + p1_ + p2_)*(1/3.0f);
 }
 
-bool plane::intersects(const ray &r) const {
-	double dummy;
+bool plane::intersects(const ray& r) const {
+	float dummy;
 	return intersects(r, dummy);
 }
 
-bool plane::intersects(const ray &r, double &t) const {
-	double numer = (p - r.pos) * norm;
-	double denom = r.dir * norm;
+bool plane::intersects(const ray& r, float& t) const {
+	float numer = (p - r.pos) * norm;
+	float denom = r.dir * norm;
 	if(denom == 0)
 		return false;
 	t = numer/denom;
